// libs/sales/product-finder/data-access/src/lib/entities/stepper/product-finder-step.enum.ts
export enum ProductFinderStepEnum {
  INTRO = 'intro',
  PREFERENCE = 'preference',
  HOUSEHOLD = 'household',
  INTERNET_USAGE = 'internetUsage',
  NUMBER_OF_SIMS = 'numberOfSims',
  MOBILE_USAGE = 'mobileUsage',
  TV_CHANNELS = 'tvChannels',
  TV_EXPERIENCE = 'tvExperience',
  RESULTS_SCREEN = 'resultsScreen',
}